gdbserver: perform signal cleanup in main application
Avoid performing exit cleanup directly in the signal handler, as these
operations are not technically async-signal-safe (the application may
have been interrupted in the middle of modifying the stlink_t, for
example).

Instead, set a flag that is checked by the application to force an exit.
Additionally, use sigaction without SA_RESTART to force syscalls to
return -EINTR when interrupted.  Thus, any blocking syscalls will return
an error after the signal is delivered, allowing us to exit.

As an aside, the GDB connection initialization is split into its own
function.